requirements.txt
# -----------------------------
# Scraping
# -----------------------------
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9

# -----------------------------
# Data & Text Processing
# -----------------------------
pandas>=2.1
tqdm>=4.65
python-dotenv>=1.0
urllib3>=2.0
restructuredtext-lint>=1.3  # optional, для перевірки тексту

# -----------------------------
# LangChain & RAG
# -----------------------------
langchain>=0.2.2
langchain-core>=0.2.2
langchain-community>=0.2.2
langchain-huggingface>=0.2.0
langchain-openai>=0.0.2

# -----------------------------
# Vector Store
# -----------------------------
chromadb>=0.4.0

# -----------------------------
# Embeddings
# -----------------------------
sentence-transformers>=2.2.2
torch>=2.1.0
transformers>=4.40.0
# -----------------------------
# UI & Visualization
# -----------------------------
streamlit>=1.30
# -----------------------------
# Testing
# -----------------------------
pytest>=7.4


langchain_groq>=1.1.1
pypdf>=3.11.5
huggingface_hub>=0.23.3
langchain-groq